Add shopinvader_product_template_multi_link_date_span
